review crud update

+++ src/review/review.controller.js
@@ -2,7 +2,7 @@ const ReviewModel = require("./review.model");
 const AnimeModel = require("../anime/anime.model");
 
 exports.create = async (req, res) => {
-  const anime = await AnimeModel.findOne({ anime_uid: req.body.uid });
+  const anime = await AnimeModel.findOne({ anime_uid: req.body.anime_uid });
   if (!anime) return res.sendStatus(404);
 
   const review = await ReviewModel.create(req.body);
@@ -13,13 +13,14 @@ exports.create = async (req, res) => {
 
 exports.delete = async (req, res) => {
   const review = await ReviewModel.findOne({ uid: req.body.uid });
-  const anime = await AnimeModel.findOne({ anime_uid: req.body.uid });
-
-  if (!review ) return res.sendStatus(404) ;
+  if (!review) return res.sendStatus(404);
+  
+  const anime = await AnimeModel.findOne({ anime_uid: req.body.anime_uid });
+  if (!anime) return res.sendStatus(404);
 
   await ReviewModel.deleteOne({ uid: req.body.uid });
   anime.reviews.pull(review);
-  await anime.save() ;
+  await anime.save();
 
   return res.sendStatus(204);
 };
@@ -31,6 +32,7 @@ exports.update = async (req, res) => {
     { uid: req.body.uid },
     req.body
   );
+
   if (updateResult.modifiedCount === 0) {
     return res.sendStatus(404); // Not Found
   }
@@ -47,12 +49,5 @@ exports.read = async (req, res) => {
   return res.status(200).json(review);
 };
 
-exports.readAllByUid = async (req, res) => {
-  const reviewList = await ReviewModel.find({ uid: req.params.uid });
 
-  if (reviewList.length === 0) {
-    return res.sendStatus(404);
-  }
 
-  return res.status(200).json(reviewList);
-};
